Re-Structuring *news* on home.js
