+++ src/app/components/jsons/news/newsBodyHomePage.css
@@ -1,0 +1,113 @@
+/* this card is inspired form this - https://georgefrancis.dev/ */
+.news-card {
+  --border-radius: 0.75rem;
+  --primary-color: #3f87da;
+  --secondary-color: #3c3852;
+  width: 100%;
+  /* max-width: fit-content; */
+  height: 220px;
+  font-family: "Arial";
+  padding: 1.2rem;
+  cursor: pointer;
+  border-radius: var(--border-radius);
+  background: #f1f1f3;
+  box-shadow: 0px 8px 16px 0px rgb(0 0 0 / 3%);
+  position: relative;
+  display: flex;
+}
+
+.news-card > * + * {
+  margin-left: 3.5em;
+}
+
+.news-card-image {
+  width: 180px;
+  height: 180px;
+  border-radius: var(--border-radius);
+  overflow: hidden;
+}
+
+.news-card .news-card-image {
+  margin-right: 25px;
+}
+.news-card .news-card-info {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: top;
+  margin-right: 25px;
+}
+.news-card .news-card__content {
+  color: var(--secondary-color);
+  font-size: 0.86rem;
+  white-space: normal;
+  text-align: justify;
+}
+
+.news-card .news-card__title {
+  padding: 0;
+  font-size: 1.3rem;
+  font-weight: bold;
+  margin-bottom: 25px;
+}
+
+.news-card .news-card__date {
+  color: #6e6b80;
+  font-size: 0.8rem;
+}
+
+.news-card .news-card__arrow {
+  position: absolute;
+  background: var(--primary-color);
+  padding: 0.4rem;
+  border-top-left-radius: var(--border-radius);
+  border-bottom-right-radius: var(--border-radius);
+  bottom: 0;
+  right: 0;
+  transition: 0.2s;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.news-card svg {
+  transition: 0.2s;
+}
+
+/* hover */
+.news-card:hover .news-card__title {
+  color: var(--primary-color);
+  text-decoration: underline;
+}
+
+.news-card:hover .news-card__arrow {
+  background: #111;
+}
+
+.news-card:hover .news-card__arrow svg {
+  transform: translateX(3px);
+}
+/* Media query's para pantallas más pequeñas que 1200px de ancho */
+
+@media (max-width: 1200px), (max-width: 1000px), (max-width: 800px) {
+  .news-card {
+    height: auto;
+  }
+}
+
+/* @media (max-width: 1200px) {
+  .news-card {
+    height: auto;
+  }
+}
+@media (max-width: 1000px) {
+  .news-card {
+    height: auto;
+  }
+}
+
+@media (max-width: 800px) {
+  .news-card {
+    height: auto;
+  }
+} */
